+++ pkg/mmath/curve.go
@@ -75,7 +75,7 @@ func (v *Curve) Normalize(begin, finish *MVec2) {
 // https://edvakf.hatenadiary.org/entry/20111016/1318716097
 // Evaluate 補間曲線を求めます。
 // return x（計算キーフレ時点のX値）, y（計算キーフレ時点のY値）, t（計算キーフレまでの変化量）
-func Evaluate(curve *Curve, start, now, end float32) (float64, float64, float64) {
+func Evaluate(curve *Curve, start, now, end int) (float64, float64, float64) {
 	if (now-start) == 0.0 || (end-start) == 0.0 {
 		return 0.0, 0.0, 0.0
 	}
@@ -135,7 +135,7 @@ func newton(x1, x2, x, t0, eps, err float64) float64 {
 }
 
 // SplitCurve 補間曲線を指定キーフレで前後に分割する
-func SplitCurve(curve *Curve, start, now, end float32) (*Curve, *Curve) {
+func SplitCurve(curve *Curve, start, now, end int) (*Curve, *Curve) {
 	if (now-start) == 0 || (end-start) == 0 {
 		return NewCurve(), NewCurve()
 	}
