+++ pkg/vmd/light_frame.go
@@ -10,7 +10,7 @@ type LightFrame struct {
 	Color      *mmath.MVec3 // 色
 }
 
-func NewLightFrame(index float32) *LightFrame {
+func NewLightFrame(index int) *LightFrame {
 	return &LightFrame{
 		BaseFrame: NewVmdBaseFrame(index),
 		Position:  mmath.NewMVec3(),
